refactor: group tests
