+++ src/test/suite/react/addUndefinedPropsToInterface.test.ts
